requirements.txt
aiogram>=3.7
aiohttp
gspread
google-auth
